Split All Servr Logic Files
