+++ mail.js
@@ -1,0 +1,12 @@
+// Module Imports
+const nodemailer = require("nodemailer");
+
+// Nodemailer Setup
+const transporter = nodemailer.createTransport({
+  host: process.env.EMAIL_HOST,
+  port: 465,
+  secure: true,
+  auth: { user: process.env.EMAIL_USER, pass: process.env.EMAIL_PASS },
+});
+
+module.exports = transporter;
