+++ public/css/adminBookings.css
@@ -172,10 +172,10 @@
 }
 
 .btn-message {
-  background-color: #3498db;
+  background-color: #2ecc71;
 }
 .btn-message:hover {
-  background-color: #21618c;
+  background-color: #27e075;
 }
 
 /* Media Queries */
