+++ statics.js
@@ -1,0 +1,26 @@
+// ===================
+// Static Data
+// ===================
+const services = [
+    { title: "VISA Assistance", description: "Fast and reliable support to process your travel visas seamlessly.", img: "services1.jpg" , link: "#contactUs"},
+    { title: "Meet and Greet Services", description: "Professional assistance to welcome and guide you through airports.", img: "services2.jpg", link: "#contactUs" },
+    { title: "Hotel Bookings", description: "Find and reserve top-quality hotels for your travels with ease.", img: "services3.jpg", link: "#contactUs" },
+    { title: "International Tickets And Local Tickets", description: "Book flights to destinations be it local or internationally.", img: "services4.jpg" , link: "/bookings"},
+    { title: "Travel and Tour Packages", description: "Discover amazing destinations with our curated travel packages.", img: "services5.jpg", link: "#contactUs" },
+    { title: "Cruises", description: "Plan your perfect getaway with our luxury cruise options.", img: "services6.jpg" , link: "#contactUs"},
+    { title: "International School Placements", description: "Helping students find the right international schools for their education.", img: "services7.jpg" , link: "#contactUs"},
+    { title: "Local And Foreign Procurements", description: "Efficient sourcing and procurement of goods and services, both locally and internationally, to meet your business and travel needs with speed and reliability.", img: "services8.jpg" , link: "#contactUs"},
+    { title: "Logistics Services", description: "Comprehensive and efficient logistics solutions designed to optimize supply chains, ensure timely deliveries, and maintain the highest standards of reliability and professionalism.", img: "services9.jpg", link: "#contactUs"}
+];
+
+const slides = [
+    { img: "slide1.jpg", title: "Discover Lagos", text: "Book your local flight today!" },
+    { img: "slide2.jpg", title: "Adventure in Paris", text: "International flights made easy." },
+    { img: "slide3.jpg", title: "Explore Qatar", text: "Luxury trips for everyone." },
+    { img: "slide4.jpg", title: "New York Awaits", text: "Fly across the globe with ease." },
+    { img: "slide5.jpg", title: "Relax in London", text: "Enjoy your perfect vacation." },
+    { img: "slide6.jpg", title: "Fly to Abuja", text: "Local or international trips, we have it all!" },
+    { img: "slide7.jpg", title: "Visit Canada", text: "Book your next adventure now!" },
+];
+
+module.exports = { services, slides };
